Moved NotifsListItemDecoration inside the adapter
